models/hovernet/requirements-hovernet.txt
absl-py==1.4.0
affine==2.3.1
antlr4-python3-runtime==4.9.3
attrs==22.2.0
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.0.4
click-plugins==1.1.1
cligj==0.7.2
config==0.5.1
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
docopt==0.6.2
future==0.18.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.48.2
hydra-core==1.3.2
idna==3.4
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
iniconfig==1.1.1
jmespath==0.10.0
joblib==1.1.1
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.0
networkx==2.5.1
numpy==1.19.1
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.3.0.36
openslide-python==1.1.2
packaging==21.3
pandas==1.1.0
Pillow==7.2.0
pluggy==1.0.0
protobuf==3.19.6
psutil==5.7.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytest==7.0.1
pytest-cov==4.0.0
pylint==2.13.9
python-dateutil==2.8.2
pytz==2023.3
pyvips==2.2.2
PyWavelets==1.1.1
PyYAML==6.0
rasterio==1.2.10
rsa==4.9
s3transfer==0.5.2
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.5.2
Shapely==1.8.5.post1
six==1.16.0
snuggs==1.4.7
tensorboard==2.3.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.1
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2020.9.3
tomli==1.2.3
torch==1.6.0
torchvision==0.7.0
tqdm==4.48.0
typing-extensions==4.1.1
urllib3==1.26.15
Werkzeug==2.0.3
zipp==3.6.0

# For FastAPI Application
fastapi==0.83.0
uvicorn==0.16.0
typing==3.7.4.3
requests==2.27.1
requests-oauthlib==1.3.1
